Move evaluation logic to helper methods
