+++ dflib/src/main/java/org/dflib/exp/num/NumberExp1.java
@@ -18,21 +18,21 @@ class NumberExp1 extends Exp1<Number, Number> implements NumExp<Number> {
 
     @Override
     public Series<Number> eval(DataFrame df) {
-        return NumberTypeResolver.eval(exp, op, df);
+        return NumberTypeEvaluator.eval(exp.eval(df), op);
     }
 
     @Override
     public Series<Number> eval(Series<?> s) {
-        return NumberTypeResolver.eval(exp, op, s);
+        return NumberTypeEvaluator.eval(exp.eval(s), op);
     }
 
     @Override
     public Number reduce(DataFrame df) {
-        return NumberTypeResolver.resolve(exp.reduce(df), op).reduce(df);
+        return NumberTypeReducer.reduce(exp.reduce(df), op);
     }
 
     @Override
     public Number reduce(Series<?> s) {
-        return NumberTypeResolver.resolve(exp.reduce(s), op).reduce(s);
+        return NumberTypeReducer.reduce(exp.reduce(s), op);
     }
 }
